review: ts_forecast accepts local_ts_csv; strict-validate train.timeseries train/split; un-skip torch-free tests
